+++ atom/model.go
@@ -1,0 +1,13 @@
+package atom
+
+import "encoding/xml"
+
+type AtomFeed struct {
+	XMLName xml.Name `xml:"feed"`
+	Title   string   `xml:"title"`
+	Entries []Entry  `xml:"entry"`
+}
+
+type Entry struct {
+	Title string `xml:"title"`
+}