+++ freshrss/freshrss.go
@@ -18,7 +18,6 @@ type FreshRSSFeedPublisher struct {
 	apiToken  string // WARNING: Do not log this value as it is a secret
 	authToken string // WARNING: Do not log this value as it is a secret
 	client    *http.Client
-	logger    *log.Logger
 }
 
 func NewFreshRSSFeedPublisher(
@@ -26,22 +25,19 @@ func NewFreshRSSFeedPublisher(
 	user string,
 	apiToken string,
 	client *http.Client,
-	logger *log.Logger,
 ) *FreshRSSFeedPublisher {
 	return &FreshRSSFeedPublisher{
 		baseUrl:  baseUrl,
 		user:     user,
 		apiToken: apiToken,
 		client:   client,
-		logger:   logger,
 	}
 }
 
 func (f *FreshRSSFeedPublisher) Authenticate() error {
-	logger := f.logger
 
 	reqUrl := fmt.Sprintf("%s/api/greader.php/accounts/ClientLogin", f.baseUrl)
-	logger.Debugf("Authenticating with FreshRSS at %s", reqUrl)
+	log.Debugf("Authenticating with FreshRSS at %s", reqUrl)
 
 	formData := url.Values{
 		"Email":  {f.user},
@@ -58,12 +54,11 @@ func (f *FreshRSSFeedPublisher) Authenticate() error {
 		return err
 	}
 
-	logger.Info("Authenticated with FreshRSS")
+	log.Info("Authenticated with FreshRSS")
 	return nil
 }
 
 func (f *FreshRSSFeedPublisher) AddFeed(feed, name, category string) error {
-	logger := f.logger
 
 	addUrl := fmt.Sprintf(
 		"%s/api/greader.php/reader/api/0/subscription/quickadd", f.baseUrl,
@@ -73,36 +68,35 @@ func (f *FreshRSSFeedPublisher) AddFeed(feed, name, category string) error {
 		"quickadd": {feed},
 	}
 
-	logger.Debugf("Adding feed to FreshRSS %s: %s", addUrl, formData.Encode())
+	log.Debugf("Adding feed to FreshRSS %s: %s", addUrl, formData.Encode())
 
 	res, err := f.doApiRequest(addUrl, []byte(formData.Encode()), true)
 	if err != nil {
 		return err
 	}
 
-	logger.Debugf("Response from FreshRSS: %s", res)
+	log.Debugf("Response from FreshRSS: %s", res)
 
 	var resData FreshRSSAddFeedResponse
 	err = json.Unmarshal(res, &resData)
 
 	if err != nil {
-		logger.Error("Unable to parse FreshRSS response", err)
+		log.Error("Unable to parse FreshRSS response", err)
 		return err
 	}
 
 	// Add the feed to the category
 	err = f.editFeed(resData.StreamId, name, category)
 	if err != nil {
-		logger.Error("Unable to add feed to category", err)
+		log.Error("Unable to add feed to category", err)
 		return err
 	}
 
-	logger.Infof("Successfully added feed %s to FreshRSS", feed)
+	log.Infof("Successfully added feed %s to FreshRSS", feed)
 	return nil
 }
 
 func (f *FreshRSSFeedPublisher) editFeed(streamId, name, category string) error {
-	logger := f.logger
 
 	addUrl := fmt.Sprintf(
 		"%s/api/greader.php/reader/api/0/subscription/edit", f.baseUrl,
@@ -115,7 +109,7 @@ func (f *FreshRSSFeedPublisher) editFeed(streamId, name, category string) error 
 		"a":  {fmt.Sprintf("user/%s/label/%s", f.user, category)},
 	}
 
-	logger.Debugf("Adding feed to FreshRSS %s: %s", addUrl, formData.Encode())
+	log.Debugf("Adding feed to FreshRSS %s: %s", addUrl, formData.Encode())
 
 	_, err := f.doApiRequest(addUrl, []byte(formData.Encode()), true)
 	if err != nil {
@@ -127,7 +121,6 @@ func (f *FreshRSSFeedPublisher) editFeed(streamId, name, category string) error 
 
 func (f *FreshRSSFeedPublisher) doApiRequest(
 	url string, payload []byte, authHeader bool) ([]byte, error) {
-	logger := f.logger
 
 	// Set headers
 	headers := map[string]string{
@@ -141,7 +134,7 @@ func (f *FreshRSSFeedPublisher) doApiRequest(
 	reader := bytes.NewReader(payload)
 	req, err := http.NewRequest("POST", url, reader)
 	if err != nil {
-		logger.Error("Unable to create request to FreshRSS", err)
+		log.Error("Unable to create request to FreshRSS", err)
 		return nil, err
 	}
 
@@ -153,19 +146,19 @@ func (f *FreshRSSFeedPublisher) doApiRequest(
 	// Make request
 	res, err := f.client.Do(req)
 	if err != nil {
-		logger.Error("Unable to make request to FreshRSS", err)
+		log.Error("Unable to make request to FreshRSS", err)
 		return nil, err
 	}
 	defer res.Body.Close()
 	data, err := io.ReadAll(res.Body)
 
 	if err != nil {
-		logger.Error("Unable to get response data from FreshRSS", err)
+		log.Error("Unable to get response data from FreshRSS", err)
 		return nil, err
 	}
 
 	if res.StatusCode != http.StatusOK && res.StatusCode != http.StatusCreated {
-		logger.Errorf("Response from FreshRSS %s", data)
+		log.Errorf("Response from FreshRSS %s", data)
 		return nil, fmt.Errorf("freshrss returned an http error code %d", res.StatusCode)
 	}
 
@@ -173,8 +166,6 @@ func (f *FreshRSSFeedPublisher) doApiRequest(
 }
 
 func (f *FreshRSSFeedPublisher) parsePlainTextAuthResponse(respData []byte) error {
-
-	logger := f.logger
 
 	lines := strings.Split(string(respData), "\n")
 	for _, line := range lines {
@@ -184,7 +175,7 @@ func (f *FreshRSSFeedPublisher) parsePlainTextAuthResponse(respData []byte) erro
 	}
 
 	if f.authToken == "" {
-		logger.Error("Unable to parse FreshRSS auth response")
+		log.Error("Unable to parse FreshRSS auth response")
 		return fmt.Errorf("unable to parse FreshRSS auth response")
 	}