+++ github/github.go
@@ -1,6 +1,7 @@
 package github
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -15,16 +16,16 @@ import (
 // user.
 type GitHubStarredFeedBuilder struct {
 	token  string // WARNING: Do not log this value as it is a secret
+	ctx    context.Context
 	client *http.Client
-	logger *log.Logger
 }
 
 func NewGitHubStarredFeedBuilder(
 	token string,
 	client *http.Client,
-	logger *log.Logger,
 ) *GitHubStarredFeedBuilder {
-	return &GitHubStarredFeedBuilder{token: token, client: client, logger: logger}
+	ctx := context.Background()
+	return &GitHubStarredFeedBuilder{token: token, ctx: ctx, client: client}
 }
 
 // This will return all starred repos including the Atom feeds for their releases
@@ -62,7 +63,6 @@ func (gh *GitHubStarredFeedBuilder) GetStarredRepos() ([]GitHubRepo, error) {
 
 // This method handles making API requests to GitHub using its REST API.
 func (gh *GitHubStarredFeedBuilder) doApiRequest(url string) (*GithubResponse, error) {
-	logger := gh.logger
 
 	headers := map[string]string{
 		"Authorization":        fmt.Sprintf("Bearer %s", gh.token),
@@ -72,9 +72,9 @@ func (gh *GitHubStarredFeedBuilder) doApiRequest(url string) (*GithubResponse, e
 		"Accept":               "application/json",
 	}
 
-	httpRequest, err := http.NewRequest("GET", url, nil)
+	httpRequest, err := http.NewRequestWithContext(gh.ctx, "GET", url, nil)
 	if err != nil {
-		logger.Error("Unable to build request to github", err)
+		log.Error("Unable to build request to github", err)
 		return nil, err
 	}
 
@@ -84,7 +84,7 @@ func (gh *GitHubStarredFeedBuilder) doApiRequest(url string) (*GithubResponse, e
 
 	res, err := gh.client.Do(httpRequest)
 	if err != nil {
-		logger.Error("Unable to make request to Github", err)
+		log.Error("Unable to make request to Github", err)
 		return nil, err
 
 	}
@@ -97,7 +97,7 @@ func (gh *GitHubStarredFeedBuilder) doApiRequest(url string) (*GithubResponse, e
 
 	ghResponse, err := gh.processGithubResponse(res)
 	if err != nil {
-		logger.Error("Unable to parse response from github", err)
+		log.Error("Unable to parse response from github", err)
 		return nil, err
 	}